+++ src/dashboard/App.tsx
@@ -18,6 +18,7 @@ import CommandPalette from '../components/CommandPalette'
 import { EditorShellLayout } from './components/EditorShellLayout'
 import { StatusBar } from './components/StatusBar'
 import { useAppearance } from '../hooks/useAppearance'
+import { OnboardingScreen } from './components/OnboardingScreen'
 import {
   HooksPage,
   PlansPage,
@@ -251,6 +252,7 @@ export default function App() {
 
   return (
     <MotionConfig reducedMotion="user">
+      <OnboardingScreen />
       {aboutOpen && <AboutDialog onClose={() => setAboutOpen(false)} />}
       {whatsNewOpen && <WhatsNewDialog onClose={() => setWhatsNewOpen(false)} />}
       <Routes>